[ahrs] float_dcm: update state interface from wrapper
